added auth validation errors

+++ app/Http/Controllers/PostController.php
@@ -2,13 +2,18 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Post;
 use Illuminate\Http\Request;
+use App\Http\Requests\PostRequest;
+use App\Models\Post;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -46,14 +51,14 @@ class PostController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(PostRequest $request)
     {
         $post = new Post();
         $post->title = $request->title;
         $post->short_title = Str::length($request->title) > 30
             ? Str::substr($request->title, 0, 30) . '...' : $request->title;
         $post->description = $request->description;
-        $post->author_id = rand(1, 4);
+        $post->author_id = \Auth::user()->id;
         if ($request->file('image')) {
             $path = Storage::putFile('public', $request->file('image'));
             $url = Storage::url($path);
@@ -73,6 +78,10 @@ class PostController extends Controller
     {
         $post = Post::join('users', 'author_id', '=', 'users.id')
             ->find($id);
+
+        if (!$post) {
+            return redirect()->route('posts.index')->withErrors('Ты куда-то не туда пытался зайти!');
+        }
         return view('posts.show', compact('post'));
     }
 
@@ -85,6 +94,12 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if (!$post) {
+            return redirect()->route('posts.index')->withErrors('Ты куда-то не туда пытался зайти!');
+        }
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('posts.index')->withErrors('Вы не можете редактировать данный пост');
+        }
         return view('posts.edit', compact('post'));
     }
 
@@ -95,9 +110,18 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+
+        if (!$post) {
+            return redirect()->route('posts.index')->withErrors('Ты куда-то не туда пытался зайти!');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('posts.index')->withErrors('Вы не можете редактировать данный пост');
+        }
+
         $post->title = $request->title;
         $post->short_title = Str::length($request->title) > 30
             ? Str::substr($request->title, 0, 30) . '...' : $request->title;
@@ -121,6 +145,14 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if (!$post) {
+            return redirect()->route('posts.index')->withErrors('Ты куда-то не туда пытался зайти!');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('posts.index')->withErrors('Вы не можете редактировать данный пост');
+        }
         $post->delete();
         return redirect()->route('posts.index')->with('success', 'Пост успешно удален!');
     }
